utils(tests): fix duplicated tests
