+++ api/utils/tests/test_web_scraping.py
@@ -1,8 +1,7 @@
 from string import ascii_letters as letters, digits
 from collections import defaultdict
 from rest_framework.test import APITestCase
-from utils import web_scraping as wbp, sessions as sns
-from datetime import datetime
+from utils import web_scraping as wbp
 from django.urls import reverse
 from pathlib import Path
 import random
@@ -72,74 +71,4 @@ class WebScrapingTest(APITestCase):
         disciplines = self.make_disciplines_request('table')
 
         self.assertFalse(len(disciplines))
-
-
-class DateUtilsTest(APITestCase):
-
-    def compare_date(self, date: datetime):
-        before_dec = datetime(date.year, 12, 30) >= date
-        after_may = datetime(date.year, 5, 1) <= date
-
-        return before_dec and after_may
-
-    def change_date(self, date: datetime):
-        month = 2 if self.compare_date(date) else 6
-        return datetime(date.year + self.compare_date(date), month=month, day=1)
-
-    def test_get_current_year_and_period(self):
-        year, period = sns.get_current_year_and_period()
-
-        date = datetime.now()
-        self.assertEqual(year, str(date.year))
-
-        self.assertEqual((period == '2'), self.compare_date(date))
-        self.assertEqual((period == '1'), not self.compare_date(date))
-
-    def test_get_next_period(self):
-        _, period = sns.get_next_period()
-
-        date = datetime.now()
-        self.assertEqual((period == '2'), not self.compare_date(date))
-        self.assertEqual((period == '1'), self.compare_date(date))
-
-        date = self.change_date(date)
-        _, period = sns.get_next_period(date)
-
-        self.assertEqual((period == '2'), not self.compare_date(date))
-        self.assertEqual((period == '1'), self.compare_date(date))
-
-    def test_get_next_period_with_year(self):
-        year, period = sns.get_next_period()
-
-        date = datetime.now()
-        self.assertEqual(str(date.year + (not (period == '2'))), year)
-
-        date = self.change_date(date)
-        year, period = sns.get_next_period(date)
-
-        self.assertEqual(str(date.year + (not (period == '2'))), year)
-
-    def test_get_previous_period(self):
-        _, period = sns.get_previous_period()
-
-        date = datetime.now()
-        self.assertEqual((period == '2'), not self.compare_date(date))
-        self.assertEqual((period == '1'), self.compare_date(date))
-
-        date = self.change_date(date)
-        _, period = sns.get_previous_period(date)
-
-        self.assertEqual((period == '2'), not self.compare_date(date))
-        self.assertEqual((period == '1'), self.compare_date(date))
-
-    def test_get_previous_period_with_year(self):
-        year, period = sns.get_previous_period()
-
-        date = datetime.now()
-        self.assertEqual(str(date.year - (period == '2')), year)
-
-        date = self.change_date(date)
-        year, period = sns.get_previous_period(date)
-
-        self.assertEqual(str(date.year - (period == '2')), year)
-
+
